Move function to correct file
